+++ R/plot.functions.R
@@ -1483,6 +1483,115 @@ plot.nma.nodesplit <- function(x, plot.type=NULL, ...) {
 
 
 
+#' @describeIn mbnma.nodesplit Plot outputs from treatment-level nodesplit MBNMA models
+#'
+#' @param x An object of `class("nodesplit")`
+#' @param plot.type A character string that can take the value of `"forest"` to plot
+#' only forest plots, `"density"` to plot only density plots, or left as `NULL` (the
+#' default) to plot both types of plot.
+#' @param ... Arguments to be sent to [ggplot2::ggplot()]
+#'
+#' @details The S3 method `plot()` on an `nodesplit` object generates either
+#' forest plots of posterior medians and 95\\% credible intervals, or density plots
+#' of posterior densities for direct and indirect evidence.
+#'
+#' @return Plots the desired graph(s) and returns an object (or list of object if
+#' `plot.type=NULL`) of `class(c("gg", "ggplot"))`
+#'
+#' @export
+plot.nodesplit <- function(x, plot.type=NULL, ...) {
+  # ... are commands to be sent to geom_histogram
+
+  # Run checks
+  argcheck <- checkmate::makeAssertCollection()
+  checkmate::assertClass(x, "nodesplit", add=argcheck)
+  checkmate::assertChoice(plot.type, choices = c("density", "forest"), null.ok=TRUE, add=argcheck)
+  #checkmate::assertLogical(facet, add=argcheck)
+  checkmate::reportAssertions(argcheck)
+
+  if (is.null(plot.type)) {
+    plot.type <- c("density", "forest")
+  }
+
+  forestdata <- x[[1]]$forest.plot$data[0,]
+  densitydata <- x[[1]]$density.plot$data[0,]
+  forestfacet <- vector()
+  densityfacet <- vector()
+  for (i in seq_along(x)) {
+    comp <- paste(x[[i]]$comparison, collapse=" vs ")
+    temp <- x[[i]]$forest.plot$data
+    forestfacet <- append(forestfacet, rep(comp, nrow(temp)))
+    forestdata <- rbind(forestdata, temp)
+
+    temp <- x[[i]]$density.plot$data
+    densityfacet <- append(densityfacet, rep(comp, nrow(temp)))
+    densitydata <- rbind(densitydata, temp)
+  }
+  forestdata$comp <- forestfacet
+  densitydata$comp <- densityfacet
+
+
+  if ("forest" %in% plot.type) {
+    forest <-
+      ggplot2::ggplot(data=forestdata, ggplot2::aes(x=forestdata$source, y=forestdata$med,
+                                                    ymin=forestdata$l95, ymax=forestdata$u95), ...) +
+      ggplot2::geom_pointrange() +
+      ggplot2::coord_flip() +  # flip coordinates (puts labels on y axis)
+      ggplot2::xlab("") + ggplot2::ylab("Treatment effect (95% CrI)") +
+      ggplot2::theme(axis.text = ggplot2::element_text(size=10),
+                     axis.title = ggplot2::element_text(size=12),
+                     title=ggplot2::element_text(size=18)) +
+      ggplot2::theme(plot.margin=ggplot2::unit(c(1,1,1,1),"cm")) +
+      ggplot2::facet_wrap(~factor(forestdata$comp)) +
+      ggplot2::theme_bw()
+  }
+  if ("density" %in% plot.type) {
+
+    density <- ggplot2::ggplot(densitydata, ggplot2::aes(x=densitydata$value,
+                                                         linetype=densitydata$Estimate, fill=densitydata$Estimate), ...) +
+      ggplot2::geom_density(alpha=0.2) +
+      ggplot2::xlab("Treatment effect (95% CrI)") +
+      ggplot2::ylab("Posterior density") +
+      ggplot2::theme(strip.text.x = ggplot2::element_text(size=12)) +
+      ggplot2::theme(axis.text = ggplot2::element_text(size=12),
+                     axis.title = ggplot2::element_text(size=14)) +
+      ggplot2::facet_wrap(~factor(densitydata$comp)) +
+      ggplot2::theme_bw() +
+      ggplot2::labs(linetype="Evidence", fill="Evidence")
+
+    # Check for size of y-axis and allow scale=free if large difference (e.g. 50x)
+    g <- ggplot2::ggplot_build(density)
+    dens <- vector()
+    for (i in seq_along(unique(g$data[[1]]$PANEL))) {
+      dens <- append(dens, max(g$data[[1]]$density[g$data[[1]]$PANEL==unique(g$data[[1]]$PANEL)[i]]))
+    }
+    if (max(dens) > (min(dens) * 50)) {
+      density <- density + ggplot2::facet_wrap(~factor(densitydata$comp), scales="free_y")
+    }
+    print(dens)
+  }
+
+  if (identical(sort(plot.type), c("density", "forest"))) {
+    graphics::plot(forest)
+    graphics::plot(density)
+    return(invisible(list(forest, density)))
+  } else {
+    if (plot.type=="forest") {
+      graphics::plot(forest)
+      return(invisible(forest))
+    } else if (plot.type=="density") {
+      graphics::plot(density)
+      return(invisible(density))
+    }
+  }
+
+}
+
+
+
+
+
+
 #' @describeIn nma.run Plot outputs from treatment-level NMA models
 #'
 #' Results can be plotted either as a single forest plot, or facetted by agent